Line the party dashboard badge up with the mobile dock
On mobile the "Powered by Music Assistant" badge sat 4px from the screen
edge while the dock below it is inset by 12px and the player card by 16px,
so nothing lined up. It now takes the player card's inset, like the settings
save button does.

The device inset is deliberately left out: the badge is positioned inside
the layout, which is padded by it already, unlike the fixed bars that add
it themselves.
